Macroprocessor changes
